+++ src/types.ts
@@ -1,0 +1,6 @@
+export type CalendarDate = Date | number
+
+export type CalendarValues = {
+  start?: CalendarDate
+  end?: CalendarDate
+}